doc: create "Why NestedText?" section
The index page had a lot of good material, but it was getting so long
that I worried people wouldn't read it.  To help with that, I moved some
of the content into a new section devoted to making the case for
*NestedText*.

This commit also adds a section on TOML to the "Alternatives" page.