Add nestedtext-ruby link in related_projects.rst
While there are still tons of code improvements and features I would
like to add, it's already usable and released to rubygems.org! All official tests are
passing.

I will contribute with my Ruby helper for the official tests too once I've
cleaned up the code with linters and so on

https://github.com/erikw/nestedtext_tests/blob/add_ruby_api/api/nestedtext_official_tests.rb